added CinemaHallPolicy and ReservationPolicy, fixed the bug that leaded to incorrect reservation denial

+++ app/Http/Controllers/Admin/CinemaHallController.php
@@ -23,6 +23,7 @@ class CinemaHallController extends Controller
      */
     public function __construct(CinemaHallService $cinemaHallService)
     {
+        $this->authorizeResource(CinemaHall::class, 'cinema_hall');
         $this->cinemaHallService = $cinemaHallService;
     }
 
@@ -33,6 +34,7 @@ class CinemaHallController extends Controller
      */
     public function index(): View
     {
+        $this->authorize('viewAny', CinemaHall::class);
         $cinemaHalls = CinemaHall::all();
         return view('admin.cinema_halls.index', compact('cinemaHalls'));
     }
@@ -44,6 +46,7 @@ class CinemaHallController extends Controller
      */
     public function create(): View
     {
+        $this->authorize('create', CinemaHall::class);
         return view('admin.cinema_halls.create');
     }
 
@@ -55,6 +58,7 @@ class CinemaHallController extends Controller
      */
     public function store(CreateRequest $request): RedirectResponse
     {
+        $this->authorize('create', CinemaHall::class);
         $data = $request->validated();
         $cinemaHall = $this->cinemaHallService->store($data);
         return redirect()->route('cinema_halls.index');
@@ -68,6 +72,7 @@ class CinemaHallController extends Controller
      */
     public function show(CinemaHall $cinemaHall): View
     {
+        $this->authorize('view', $cinemaHall);
         $filmSessions = FilmSession::paginate(10);
         return view('admin.cinema_halls.show', compact('cinemaHall', 'filmSessions'));
     }
@@ -80,18 +85,20 @@ class CinemaHallController extends Controller
      */
     public function edit(CinemaHall $cinemaHall): View
     {
+        $this->authorize('update', $cinemaHall);
         return view('admin.cinema_halls.edit', compact('cinemaHall'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\Http\Requests\Seat\UpdateRequest  $request
      * @param  \App\Models\CinemaHall  $cinemaHall
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
-    public function update(UpdateRequest $request, CinemaHall $cinemaHall)
+    public function update(UpdateRequest $request, CinemaHall $cinemaHall): RedirectResponse
     {
+        $this->authorize('update', $cinemaHall);
         $data = $request->validated();
         $cinemaHall = $this->cinemaHallService->update($data);
         return redirect()->route('cinema_halls.show', ['cinema_hall' => $request->cinema_hall_id]);
@@ -105,6 +112,7 @@ class CinemaHallController extends Controller
      */
     public function delete(CinemaHall $cinemaHall): View
     {
+        $this->authorize('delete', $cinemaHall);
         return view('admin.cinema_halls.delete', compact('cinemaHall'));
     }
 
@@ -114,8 +122,9 @@ class CinemaHallController extends Controller
      * @param  \App\Models\CinemaHall  $cinemaHall
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function destroy(CinemaHall $cinemaHall)
+    public function destroy(CinemaHall $cinemaHall): RedirectResponse
     {
+        $this->authorize('delete', $cinemaHall);
         $film = $this->cinemaHallService->destroy($cinemaHall->id);
         return redirect()->route('cinema_halls.index');
     }
